+++ src/BattleActor.h
@@ -2,11 +2,11 @@
 
 typedef struct BattleActor BattleActor;
 	
-uint16_t BattleActor_GetSpeed(BattleActor *actor);
+int BattleActor_GetSpeed(BattleActor *actor);
 bool BattleActor_IsPlayer(BattleActor *actor);
 
-BattleActor *InitBattleActor(bool isPlayer, uint16_t level, uint16_t speed, uint16_t maxHealth);
-void DealDamage(uint16_t potency, BattleActor *defender);
+BattleActor *InitBattleActor(bool isPlayer, int level, int speed, int maxHealth);
+void DealDamage(int potency, BattleActor *defender);
 
-uint16_t BattleActor_GetHealth(BattleActor *actor);
-uint16_t BattleActor_GetMaxHealth(BattleActor *actor);
+int BattleActor_GetHealth(BattleActor *actor);
+int BattleActor_GetMaxHealth(BattleActor *actor);
